requirements.txt
falcon
pyjwt
bcrypt
rethinkdb
green
email_validator
ddt
telepot
jinja2
lxml

